# tests/IncidentCompass.IntegrationTests/Baselines/openapi-v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "IncidentCompass.Api | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://localhost/"
    }
  ],
  "paths": {
...
    "/api/v1/incidents": {
      "post": {
        "tags": [
          "v1"
        ],
        "summary": "Ingest an incident signal envelope and resolve it to a fault and triage job.",
        "operationId": "IngestIncidentSignal",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IncidentEnvelopeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestSignalResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "429": {
            "description": "Too Many Requests"
          }
        },
        "security": [
          {
            "IncidentCompassApiKey": []
          }
        ]
      }
    },
    "/api/v1/faults/{id}": {
      "get": {
        "tags": [
          "v1"
        ],
        "summary": "Return fault details, including its current triage job summary.",
        "operationId": "GetFaultById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FaultDetailsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "429": {
            "description": "Too Many Requests"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "IncidentCompassApiKey": []
          }
        ]
      }
    },
    "/api/v1/faults/{id}/ledger": {
      "get": {
        "tags": [
          "v1"
        ],
        "summary": "Return DB-ordered governed investigation ledger events for a fault.",
        "operationId": "GetFaultLedgerByFaultId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FaultLedgerResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "429": {
            "description": "Too Many Requests"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "IncidentCompassApiKey": []
          }
        ]
      }
    },
...
    "/api/v1/action-approvals": {
      "get": {
        "tags": [
          "action-approvals"
        ],
        "summary": "List tenant-scoped action approvals with paired external-resource filtering.",
        "operationId": "ListActionApprovals",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "externalResourceKind",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "externalResourceId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionApprovalListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "429": {
            "description": "Too Many Requests"
          }
        },
        "security": [
          {
            "IncidentCompassApiKey": []
          }
        ]
      }
    },
    "/api/v1/action-approvals/{id}": {
      "get": {
        "tags": [
          "action-approvals"
        ],
        "summary": "Return the immutable action proposal tuple and grounded provenance for review.",
        "operationId": "GetActionApproval",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionApprovalDetailsResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "429": {
            "description": "Too Many Requests"
          }
        },
        "security": [
          {
            "IncidentCompassApiKey": []
          }
        ]
      }
    },
    "/api/v1/action-approvals/{id}/approve": {
      "post": {
        "tags": [
          "action-approvals"
        ],
        "summary": "Approve the exact observed action payload and approval hashes.",
        "operationId": "ApproveAction",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionApprovalDetailsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "429": {
            "description": "Too Many Requests"
          }
        },
        "security": [
          {
            "IncidentCompassApiKey": []
          }
        ]
      }
    },
    "/api/v1/action-approvals/{id}/reject": {
      "post": {
        "tags": [
          "action-approvals"
        ],
        "summary": "Reject the exact observed action payload and approval hashes.",
        "operationId": "RejectAction",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionApprovalDetailsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "429": {
            "description": "Too Many Requests"
          }
        },
        "security": [
          {
            "IncidentCompassApiKey": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "ActionApprovalDetailsResponse": {
        "required": [
          "id",
          "status",
          "approvalContractVersion",
          "originReportId",
          "toolId",
          "category",
          "mode",
          "logicalTargetId",
          "canonicalPayload",
          "payloadSha256",
          "provenanceSha256",
          "approvalSha256",
          "reviewSummary",
          "createdAtUtc",
          "expiresAtUtc",
          "decisionAtUtc",
          "rejectionReason",
          "completedAtUtc",
          "resultSummary",
          "failureCode",
          "externalResourceKind",
          "externalResourceId",
          "externalBeforeState",
          "externalAfterState",
          "provenance"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string"
          },
          "approvalContractVersion": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "originReportId": {
            "type": "string",
            "format": "uuid"
          },
          "toolId": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "mode": {
            "type": "string"
          },
          "logicalTargetId": {
            "type": "string"
          },
          "canonicalPayload": {
            "$ref": "#/components/schemas/JsonElement"
          },
          "payloadSha256": {
            "type": "string"
          },
          "provenanceSha256": {
            "type": "string"
          },
          "approvalSha256": {
            "type": "string"
          },
          "reviewSummary": {
            "type": "string"
          },
          "createdAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "expiresAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "decisionAtUtc": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "rejectionReason": {
            "type": [
              "null",
              "string"
            ]
          },
          "completedAtUtc": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "resultSummary": {
            "type": [
              "null",
              "string"
            ]
          },
          "failureCode": {
            "type": [
              "null",
              "string"
            ]
          },
          "externalResourceKind": {
            "type": [
              "null",
              "string"
            ]
          },
          "externalResourceId": {
            "type": [
              "null",
              "string"
            ]
          },
          "externalBeforeState": {
            "type": [
              "null",
              "string"
            ]
          },
          "externalAfterState": {
            "type": [
              "null",
              "string"
            ]
          },
          "provenance": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ActionApprovalProvenanceResponse"
            }
          }
        }
      },
...
      "ActionApprovalListResponse": {
        "required": [
          "actions",
          "nextCursor"
        ],
        "type": "object",
        "properties": {
          "actions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ActionApprovalListItemResponse"
            }
          },
          "nextCursor": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "ApproveActionRequest": {
        "required": [
          "payloadSha256",
          "approvalSha256"
        ],
        "type": "object",
        "properties": {
          "payloadSha256": {
            "type": "string"
          },
          "approvalSha256": {
            "type": "string"
          }
        }
      },
...
      "FaultDetailsResponse": {
        "required": [
          "id",
          "status",
          "fingerprint",
          "fingerprintVersion",
          "fingerprintStrength",
          "canGroup",
          "serviceName",
          "environment",
          "severity",
          "correlationId",
          "triggerSignalId",
          "recurrenceOf",
          "createdAtUtc",
          "completedAtUtc",
          "job"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string"
          },
          "fingerprint": {
            "type": "string"
          },
          "fingerprintVersion": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "fingerprintStrength": {
            "type": "string"
          },
          "canGroup": {
            "type": "boolean"
          },
          "serviceName": {
            "type": "string"
          },
          "environment": {
            "type": "string"
          },
          "severity": {
            "type": [
              "null",
              "string"
            ]
          },
          "correlationId": {
            "type": [
              "null",
              "string"
            ]
          },
          "triggerSignalId": {
            "type": "string",
            "format": "uuid"
          },
          "recurrenceOf": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "createdAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "completedAtUtc": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "job": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TriageJobSummary"
              }
            ]
          }
        }
      },
      "FaultLedgerEventResponse": {
        "required": [
          "id",
          "jobId",
          "attempt",
          "eventType",
          "role",
          "toolName",
          "decision",
          "rationale",
          "decisionReason",
          "toolStatus",
          "tokensDelta",
          "workersDelta",
          "payloadRef",
          "configHash",
          "createdAtUtc"
        ],
        "type": "object",
        "properties": {
          "id": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "jobId": {
            "type": "string",
            "format": "uuid"
          },
          "attempt": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "eventType": {
            "type": "string"
          },
          "role": {
            "type": [
              "null",
              "string"
            ]
          },
          "toolName": {
            "type": [
              "null",
              "string"
            ]
          },
          "decision": {
            "type": [
              "null",
              "string"
            ]
          },
          "rationale": {
            "type": [
              "null",
              "string"
            ]
          },
          "decisionReason": {
            "type": [
              "null",
              "string"
            ]
          },
          "toolStatus": {
            "type": [
              "null",
              "string"
            ]
          },
          "tokensDelta": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "workersDelta": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "payloadRef": {
            "type": [
              "null",
              "string"
            ]
          },
          "configHash": {
            "type": "string"
          },
          "createdAtUtc": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "FaultLedgerResponse": {
        "required": [
          "faultId",
          "events"
        ],
        "type": "object",
        "properties": {
          "faultId": {
            "type": "string",
            "format": "uuid"
          },
          "events": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FaultLedgerEventResponse"
            }
          }
        }
      },
...
      "IncidentCorrelationRequest": {
        "required": [
          "traceId",
          "spanId",
          "parentSpanId",
          "externalId"
        ],
        "type": "object",
        "properties": {
          "traceId": {
            "type": [
              "null",
              "string"
            ]
          },
          "spanId": {
            "type": [
              "null",
              "string"
            ]
          },
          "parentSpanId": {
            "type": [
              "null",
              "string"
            ]
          },
          "externalId": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "IncidentEnvelopeRequest": {
        "required": [
          "sourceKind",
          "serviceName",
          "environment",
          "severity",
          "summary",
          "description",
          "observedAtUtc",
          "correlation",
          "attributes",
          "payload"
        ],
        "type": "object",
        "properties": {
          "sourceKind": {
            "type": "string"
          },
          "serviceName": {
            "type": [
              "null",
              "string"
            ]
          },
          "environment": {
            "type": [
              "null",
              "string"
            ]
          },
          "severity": {
            "type": [
              "null",
              "string"
            ]
          },
          "summary": {
            "type": [
              "null",
              "string"
            ]
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "observedAtUtc": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "correlation": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/IncidentCorrelationRequest"
              }
            ]
          },
          "attributes": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/JsonNode"
              }
            ]
          },
          "payload": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/JsonNode"
              }
            ]
          }
        }
      },
      "IngestSignalResponse": {
        "required": [
          "signalId",
          "faultId",
          "isNewFault",
          "isNewJob",
          "isSuppressed",
          "jobId",
          "configHash"
        ],
        "type": "object",
        "properties": {
          "signalId": {
            "type": "string",
            "format": "uuid"
          },
          "faultId": {
            "type": "string",
            "format": "uuid"
          },
          "isNewFault": {
            "type": "boolean"
          },
          "isNewJob": {
            "type": "boolean"
          },
          "isSuppressed": {
            "type": "boolean"
          },
          "jobId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "configHash": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "JsonElement": {},
      "JsonNode": {},
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "RejectActionRequest": {
        "required": [
          "payloadSha256",
          "approvalSha256",
          "reason"
        ],
        "type": "object",
        "properties": {
          "payloadSha256": {
            "type": "string"
          },
          "approvalSha256": {
            "type": "string"
          },
          "reason": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "TriageJobSummary": {
        "required": [
          "id",
          "status",
          "attempt",
          "configHash",
          "createdAtUtc"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string"
          },
          "attempt": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "configHash": {
            "type": "string"
          },
          "createdAtUtc": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    },
    "securitySchemes": {
      "IncidentCompassApiKey": {
        "type": "apiKey",
        "description": "Host-issued IncidentCompass API key.",
        "name": "X-IncidentCompass-Key",
        "in": "header"
      }
    }
  },
...
}